Split SpatialJoinAlgorithms into one class per algorithm
The monolithic SpatialJoinAlgorithms class mixed five unrelated join
algorithms (baseline, S2 point, S2 point/polyline, bounding-box r-tree,
libspatialjoin) together with their shared and not-so-shared helpers.
Split it into src/engine/spatialJoinAlgorithms/: an abstract
SpatialJoinAlgorithms base with only what every algorithm needs, an
intermediate RtreeEntryAlgorithm base for the two algorithms that parse
arbitrary geometries into RtreeEntrys (Baseline, BoundingBox), one file
per concrete algorithm implementing run(), and a SpatialJoinGeoUtils
free-function module for the stateless helpers used outside of any
algorithm instance (index parsing, materialized-view geometry index).

No behavior change; algorithm bodies are unchanged aside from moving
between files/classes.
