+++ src/engine/CMakeLists.txt
@@ -13,7 +13,12 @@ add_library(engine
         VariableToColumnMap.cpp ExportQueryExecutionTrees.cpp
         CartesianProductJoin.cpp TextIndexScanForWord.cpp TextIndexScanForEntity.cpp
         TextLimit.cpp LazyGroupBy.cpp GroupByHashMapOptimization.cpp SpatialJoin.cpp
-        CountConnectedSubgraphs.cpp SpatialJoinAlgorithms.cpp PathSearch.cpp ExecuteUpdate.cpp
+        CountConnectedSubgraphs.cpp
+        spatialJoinAlgorithms/SpatialJoinAlgorithms.cpp spatialJoinAlgorithms/SpatialJoinGeoUtils.cpp
+        spatialJoinAlgorithms/RtreeEntryAlgorithm.cpp
+        spatialJoinAlgorithms/BaselineAlgorithm.cpp spatialJoinAlgorithms/S2GeometryAlgorithm.cpp
+        spatialJoinAlgorithms/S2PointPolylineAlgorithm.cpp spatialJoinAlgorithms/BoundingBoxAlgorithm.cpp
+        spatialJoinAlgorithms/LibspatialjoinAlgorithm.cpp PathSearch.cpp ExecuteUpdate.cpp
         Describe.cpp GraphStoreProtocol.cpp SpatialJoinParser.cpp SpatialJoinCachedIndex.cpp
         QueryExecutionContext.cpp ExistsJoin.cpp SparqlProtocol.cpp ParsedRequestBuilder.cpp
         NeutralOptional.cpp Load.cpp StripColumns.cpp NamedResultCache.cpp
