+++ src/engine/spatialJoinAlgorithms/S2GeometryAlgorithm.cpp
@@ -1,0 +1,91 @@
+// Copyright 2024 - 2026 The QLever Authors, in particular:
+//
+// 2024 - 2025 Jonathan Zeller github@Jonathan24680, UFR
+// 2024 - 2026 Christoph Ullinger <ullingec@informatik.uni-freiburg.de>, UFR
+//
+// UFR = University of Freiburg, Chair of Algorithms and Data Structures
+
+// You may not use this file except in compliance with the Apache 2.0 License,
+// which can be found in the `LICENSE` file at the root of the QLever project.
+
+#include "engine/spatialJoinAlgorithms/S2GeometryAlgorithm.h"
+
+#include <s2/s2closest_point_query.h>
+#include <s2/s2earth.h>
+#include <s2/s2point.h>
+#include <s2/s2point_index.h>
+#include <s2/util/units/length-units.h>
+
+#include "engine/spatialJoinAlgorithms/SpatialJoinGeoUtils.h"
+#include "util/GeoConverters.h"
+
+using namespace geometryConverters;
+
+// ____________________________________________________________________________
+Result S2GeometryAlgorithm::run() {
+  const auto [idTableLeft, resultLeft, idTableRight, resultRight, leftJoinCol,
+              rightJoinCol, leftSelectedCols, rightSelectedCols, numColumns,
+              maxDist, maxResults, joinType, de9imFilter, rightCacheName,
+              bbLeft, bbRight] = params_;
+  IdTable result{numColumns, qec_->getAllocator()};
+
+  S2PointIndex<size_t> s2index;
+
+  // Optimization: If we only search by maximum distance, the operation is
+  // symmetric, so the larger table can be used for the index
+  bool indexOfRight =
+      (maxResults.has_value() || (idTableLeft->size() > idTableRight->size()));
+  auto indexTable = indexOfRight ? idTableRight : idTableLeft;
+  auto indexJoinCol = indexOfRight ? rightJoinCol : leftJoinCol;
+
+  // Populate the index
+  for (size_t row = 0; row < indexTable->size(); row++) {
+    auto p = ad_utility::detail::spatialjoin::getPoint(indexTable, row,
+                                                       indexJoinCol);
+    if (p.has_value()) {
+      s2index.Add(toS2Point(p.value()), row);
+    }
+  }
+  // Performs a nearest neighbor search on the index and returns the closest
+  // points that satisfy the criteria given by `maxDist_` and `maxResults_`.
+
+  // Construct a query object with the given constraints
+  auto s2query = S2ClosestPointQuery<size_t>{&s2index};
+
+  if (maxResults.has_value()) {
+    s2query.mutable_options()->set_max_results(
+        static_cast<int>(maxResults.value()));
+  }
+  if (maxDist.has_value()) {
+    s2query.mutable_options()->set_inclusive_max_distance(S2Earth::ToAngle(
+        util::units::Meters(static_cast<float>(maxDist.value()))));
+  }
+
+  auto searchTable = indexOfRight ? idTableLeft : idTableRight;
+  auto searchJoinCol = indexOfRight ? leftJoinCol : rightJoinCol;
+  // Use the index to lookup the points of the other table
+  for (size_t searchRow = 0; searchRow < searchTable->size(); searchRow++) {
+    auto p = ad_utility::detail::spatialjoin::getPoint(searchTable, searchRow,
+                                                       searchJoinCol);
+    if (!p.has_value()) {
+      continue;
+    }
+    auto s2target =
+        S2ClosestPointQuery<size_t>::PointTarget{toS2Point(p.value())};
+
+    for (const auto& neighbor : s2query.FindClosestPoints(&s2target)) {
+      // In this loop we only receive points that already satisfy the given
+      // criteria
+      auto indexRow = neighbor.data();
+      auto dist = S2Earth::ToKm(neighbor.distance());
+
+      auto rowLeft = indexOfRight ? searchRow : indexRow;
+      auto rowRight = indexOfRight ? indexRow : searchRow;
+      addResultTableEntry(&result, idTableLeft, idTableRight, rowLeft, rowRight,
+                          Id::makeFromDouble(dist));
+    }
+  }
+
+  return Result(std::move(result), std::vector<ColumnIndex>{},
+                Result::getMergedLocalVocab(*resultLeft, *resultRight));
+}
