+++ src/global/RuntimeParameters.h
@@ -157,7 +157,7 @@ struct RuntimeParameters {
   // prefilter-free baseline, or for debugging, as wrong results may be
   // related to the `PrefilterExpression`s.
   Bool enablePrefilterOnIndexScans_{true, "enable-prefilter-on-index-scans"};
-  // The maximum number of threads to be used in `SpatialJoinAlgorithms`.
+  // The maximum number of threads to be used by the spatial join algorithms.
   SizeT spatialJoinMaxNumThreads_{8, "spatial-join-max-num-threads"};
   // The maximum number of threads for the parallel counting loops of the
   // pattern trick (see `CountAvailablePredicates`). The value `0` means the
@@ -172,7 +172,7 @@ struct RuntimeParameters {
   // returns for more threads.
   SizeT parallelSortNumThreads_{3, "parallel-sort-num-threads"};
   // The maximum size of the `prefilterBox` for
-  // `SpatialJoinAlgorithms::libspatialjoinParse()`.
+  // `LibspatialjoinAlgorithm::libspatialjoinParse()`.
   SizeT spatialJoinPrefilterMaxSize_{2'500, "spatial-join-prefilter-max-size"};
   // Push joins into both children of unions if this leads to a cheaper
   // cost-estimate.
