+++ test/engine/SpatialJoinPrefilterTestHelpers.h
@@ -24,8 +24,9 @@
 #include "engine/QueryExecutionContext.h"
 #include "engine/QueryExecutionTree.h"
 #include "engine/SpatialJoin.h"
-#include "engine/SpatialJoinAlgorithms.h"
 #include "engine/SpatialJoinConfig.h"
+#include "engine/spatialJoinAlgorithms/LibspatialjoinAlgorithm.h"
+#include "engine/spatialJoinAlgorithms/SpatialJoinGeoUtils.h"
 #include "global/RuntimeParameters.h"
 #include "rdfTypes/GeoSparqlHelpers.h"
 #include "rdfTypes/GeometryInfo.h"
@@ -191,7 +192,7 @@ inline sj::SweeperCfg makeSweeperCfg(const LibSpatialJoinConfig& libSJConfig,
                                      double withinDist) {
   using enum SpatialJoinType::Enum;
   sj::SweeperCfg cfg =
-      SpatialJoinAlgorithms::libspatialjoinSweeperConfig(1, 1_GB);
+      ad_utility::detail::spatialjoin::libspatialjoinSweeperConfig(1, 1_GB);
   cfg.withinDist = withinDist;
   auto joinTypeVal = libSJConfig.joinType_;
   cfg.writeRelCb = [&results, &resultDists, joinTypeVal](
@@ -243,14 +244,14 @@ inline void runParsingAndSweeper(
   std::shared_ptr<Operation> op = spatialJoinOperation->getRootOperation();
   SpatialJoin* spatialJoin = static_cast<SpatialJoin*>(op.get());
 
-  // Build `SpatialJoinAlgorithms` instance from spatial join operation
+  // Build `LibspatialjoinAlgorithm` instance from spatial join operation
   auto prepared = spatialJoin->onlyForTestingGetPrepareJoin();
-  SpatialJoinAlgorithms sjAlgo{qec, prepared, config, spatialJoin};
+  LibspatialjoinAlgorithm sjAlgo{qec, prepared, config, spatialJoin};
 
   // The regular implementation can also be tested instead of this mock version,
   // but then only limited information is available.
   if (useRegularImplementation) {
-    auto result = sjAlgo.LibspatialjoinAlgorithm();
+    auto result = sjAlgo.run();
     auto varToCol = spatialJoin->computeVariableToColumnMap();
     auto leftCol = varToCol.at(varLeft).columnIndex_;
     auto rightCol = varToCol.at(varRight).columnIndex_;
